feat: Move DTO AssetMergeInfo
By moving the DTO to a shared lib it can now be reused in other parts of
UVMS to deserialize the event payload.

Refs: FART-530
