+++ module/src/main/java/fish/focus/uvms/rest/asset/service/SSEResource.java
@@ -1,11 +1,11 @@
 package fish.focus.uvms.rest.asset.service;
 
+import fish.focus.uvms.asset.domain.entity.Asset;
+import fish.focus.uvms.asset.message.event.UpdatedAssetEvent;
+import fish.focus.uvms.asset.remote.dto.AssetMergeInfo;
 import fish.focus.uvms.commons.date.JsonBConfigurator;
 import fish.focus.uvms.rest.security.RequiresFeature;
 import fish.focus.uvms.rest.security.UnionVMSFeature;
-import fish.focus.uvms.asset.domain.entity.Asset;
-import fish.focus.uvms.asset.dto.AssetMergeInfo;
-import fish.focus.uvms.asset.message.event.UpdatedAssetEvent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
