+++ module/src/test/java/fish/focus/uvms/asset/message/EventStreamSenderTest.java
@@ -5,12 +5,12 @@
  */
 package fish.focus.uvms.asset.message;
 
-import fish.focus.uvms.commons.date.JsonBConfigurator;
-import fish.focus.uvms.commons.message.api.MessageConstants;
 import fish.focus.uvms.asset.domain.dao.AssetDao;
 import fish.focus.uvms.asset.domain.entity.Asset;
 import fish.focus.uvms.asset.domain.entity.AssetRemapMapping;
-import fish.focus.uvms.asset.dto.AssetMergeInfo;
+import fish.focus.uvms.asset.remote.dto.AssetMergeInfo;
+import fish.focus.uvms.commons.date.JsonBConfigurator;
+import fish.focus.uvms.commons.message.api.MessageConstants;
 import fish.focus.uvms.mobileterminal.timer.AssetRemapTask;
 import fish.focus.uvms.tests.BuildAssetServiceDeployment;
 import org.jboss.arquillian.container.test.api.OperateOnDeployment;
